MINOR Add elementary page with title and canvas
This is just a simple commit that has a simple index.php page with the
game title and a canvas set up. It also has two external file linked
(both are JavaScript and CSS files).

The page exhibits the title "MNEMO" and a canvas with border (debug).
The page also outputs in console if the JavaScript code was executed.