MINOR Add basic game table with cards fliped down
Dropped HTML Canvas. Everything will be loaded images within DIV
elements.

The game design in its basic form was made. The HTML structure and image
files for a basic memory game - which, in this case, are numbers in
arabic and japanese.

The CSS file made the smallest changes to make the basic visual (can be
changed). It also blocks selection to not undermine the game.

The JavaScript file names was changed. It was done so to reflect the
fact the game won't be in HTML Canvas anymore.

The basic game strucutre includes a score, measures for cards and
distances, list of images for cards and a list of cards (ordered cards,
shuffled cards, active cards (clicked and open) and those that were
paired). There are also helper functions and the start game function -
which is called after everything is loaded.

The code for start() resets variables at the beggining to make sure
everything is starting at the point. Then, it sets the game 'table' with
the shuffled cards (todo) facing down.