MINOR Add support for flipping a pair of cards
This adds support for flipping cards. The number of cards able to be
flipped at a given time is two. Note that "flipping" means to make the
clicked card either face up or down - depends on its previous state.

This commit also includes a randomizer of cards based on the
Fisher-Yates algorithm. To be specific, it uses the revised version by
Durstenfeld that applies it to computers.

The code also:

- set/get the state of the card (face down/face up);
- set/get active cards (cards faced up);
- get paired cards (not usable because there is no set function);
- avoids copy by reference with non-primitive types (throw/catch);
- forces copy by reference with var declaration to search non-primitives
  inside arrays with indexOf;
- sets click event for every card;