MINOR Add auto-unflip, score and cardmatch systems
This commit adds the auto-unflip functionality that unflip cards not
paired back to being face down. Also adds +5 for made pairs and -2 for
wrong pairing. Finally, can detect if cards are a match.

If two cards are faced up, the code checks if they are a pair. If they
are not, it decreses the score (current and total) and flip those cards
to face down. If they are, increases the score (current and total), let
the cards faced up and store those cards into the paired cards list. In
both scenarios, the active cards list is cleaned and all tracker flags
are properly set.

The code also:
- Verify if the cards are paired (function)
- Verify if the active cards are numbering two
- Set paired cards into a proper list (function)

TODO Make the auto-unflip slow enough for the cards to be seen and
looked for by the player. Currently, the second card can't be seen
unless it matches another card and make a pair.