MINOR Add lock and timing system
The lock system prevents code from being executed. Especifically, the
code that deals with clicks, making sure cards don't instant flip or
that user misclick and potentially breaks the code (super human speed
needed but one never knows).

The timing system holds the cards facing up for as long as needed for a
user to read and memorize them. Only after, in case they are not paired
cards, those cards are flipped to be facing down.