MINOR Fix problem with non-deleted HTMl positions
In cards.active, html_positions was not being deleted upon a restart of
the game. If any garbage was left there, problems would arise.

If a player click on a card and never on a second card and resets the
game, that card would been left there. That is an unpredicted behavior
that was now fixed.