MINOR Refactor set_current_score function name
This function name wasn't reflecting what it does: updates the score
vale and draws it on the screen. The change reflects this.

The function name now is update_score.

TODO set_score to set score values
TODO Needs to be refactored again with set_score being used to set score
values