MINOR Add reset button to start everything from 0
There is now a New Game button (to start a new game inside a series of
games) and a Reset button (to start a whole new series of games).

The New Game button changes current score to zero. The Reset button
changes both current and total scores to zero.

The reset function encapsulates the start function together with a
statement to zero the score.total variable. The calls are made first
with score.total = 0 followed by a call to start function.